Nack three CVEs in busybox.
These all require FEATURE_CLEANUP to be set at compile time, which we don't set.
Ref: https://github.com/wolfi-dev/os/blob/main/busybox/busyboxconfig#L36

Busybox upstream does not depend on free() logic to clean up, it uses exit() becuase it is
designed to be short-lived. The non-default FEATURE_CLEAN_UP setting does trigger free() calls,
but we don't compile with that.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>