apache-nifi: update advisory for GHSA-rc42-6c7j-7h5r
We have tried to bump boot spring via pombump and later by patching the
related pom.xml files to version 3.4.5, but unfortunately they all
resulted in build failures.

Upstream has currently a commit in main and in their 2.4.0-RC1 branch,
which means that the bump will be available in apache-nifi v2.4.0.
https://github.com/apache/nifi/commit/954c9c4bdfc174b4af9a73789b5bdea896d9bd69

We now need to wait for a new release, or for upstream to backport the
fix into their 2.3.0 branch.

Signed-off-by: David Negreira <david.negreira@chainguard.dev>