airflow-3: update advisory (#20973)
Update advisory for CVE-2025-48924
The org.apache.commons:commons-lang3 is pulled through a transient
dependency of a python package named 'ray' version 2.47.1 which is the
latest. We will have to wait for 'ray' to push a more recent version of
commons-lang3 and airflow to push a new release with the new 'ray'
dependency updated.