juicefs-1.3: update advisories
Update advisory for GHSA-wf43-55jj-vwq8
This package must be removed from upstream dependencies. Upstream
already consumes the fixed version, and trying to bump the vulnerable
version will cause build failures due to duplication.

Update advisory for GHSA-cvx7-x8pj-x2gw
This dependency is brought in via a transient dependency. Any attempts
to bump to a fixed version of this dependency result in a build failure,
hence we will need upstream to update their dependencies in order to fix
this issue.

Signed-off-by: David Negreira <david.negreira@chainguard.dev>