grafana-pyroscope-1.13: update advisory (#21334)
Update advisory for GHSA-7rh7-c77v-6434
The oauth2-proxy is a transient dependency and any attempts to bump
result in build failure. We will have to wait for upstream to work on
bumping their dependency tree. There is currently an issue open upstream
to try to bump this dependency but it is still a work in progress:
https://github.com/grafana/pyroscope/pull/4335

Signed-off-by: David Negreira <david.negreira@chainguard.dev>