apache-nifi-registry: update advisories
Advisory for GHSA-fghv-69vj-qj49
The fix for this CVE requires upgrading netty-codec-http to 4.2.5.Final.
However, this dependency is bundled as part of the AWS extension JAR and
cannot be independently updated through dependency management. We need
to wait for Apache NiFi 2.6.0 to be officially released before we can
apply this security fix.

Advisory for GHSA-3p8m-j85q-pgmj
The fix for this CVE requires upgrading netty-codec-compression to
4.2.5.Final. However, this dependency is bundled as part of the AWS
extension JAR and cannot be independently updated through dependency
management. We need to wait for Apache NiFi 2.6.0 to be officially
released before we can apply this security fix.

Signed-off-by: David Negreira <david.negreira@chainguard.dev>