selenium: update advisories
Update advisories GHSA-3p8m-j85q-pgmj and GHSA-fghv-69vj-qj49
We are unable to bump netty-codec or netty-codec-http to a newer version
without breaking the build. Upstream already has a fix in place for a
future version, and we should await for a new build with the fixed
versions. More information:
https://github.com/SeleniumHQ/selenium/pull/16194

Signed-off-by: David Negreira <david.negreira@chainguard.dev>