akhq/celeborn: update advisories (#23399)
Update advisory of akhq for GHSA-3p8m-j85q-pgmj
The netty-codec-compression dependency is a transient dependency from
Micronaut, we will need to wait for Micronaut to release a fixed version
and akhq to upgrade their dependencies to use the new Micronaut version.

Update advisory of celeborn for GHSA-3p8m-j85q-pgmj and GHSA-fghv-69vj-qj49
This dependency is a transient dependency from the apache
ratis-thirdparty-misc plugins, we will need to wait for a new version of
ratis-thirdparty-misc to release a fixed version and celeborn to upgrade
their dependencies to use the new ratis-thirdparty-misc version.

Signed-off-by: David Negreira <david.negreira@chainguard.dev>