py3-pip: update advisory
Update advisory for CVE-2025-8869
Unfortunately we do not have a way to express in scanners and in our
systems that a python package has cherry-picks from upstream that fixes
vulnerabilities in python. So we will have for upstream to release pip
25.3 so that we have a way for our system and scanners to ensure that a
patched version is present.

This is the PR where we have patched py3-pip with the upstream patches:
https://github.com/wolfi-dev/os/pull/67721