dotnet-bootstrap-{8,9}: update advisory
Update advisory for CVE-2025-55315
The dotnet-bootstrap package is used to bootstrap the dotnet package.
The baseline version in .NET servicing builds references the previous release for API validation only.
It ensures compatibility and does not affect shipped runtime or package contents.
Although 8.0.20 version of dotnet is listed in the dotnet-boostrap package, the build used in the dotnet package actually produces and ships 8.0.21 which includes the CVE fix.

More information about the baseline compatibility can be found here:
https://learn.microsoft.com/en-us/dotnet/fundamentals/apicompat/package-validation/baseline-version-validator
as well as in the code
https://github.com/dotnet/sdk/blob/0f5c106d6a6eb9da4dd4adb2a28bd4024ba9287a/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.ApiCompat.targets#L22

Signed-off-by: David Negreira <david.negreira@chainguard.dev>