airflow-3: update advisories
Update advisories for GHSA-hrfv-mqp8-q5rw, GHSA-f9vj-2wh5-fj8j,
GHSA-2g68-c3qc-8985 and GHSA-q34m-jh98-gwm2

We are currently unable to bump the Werkzeug dependency as there are
some hard constraints on the connnexion dependency for the FAB auth
manager APIs.

More information can be found in a comment by upstream maintainers here:
https://github.com/apache/airflow/pull/51681\#issuecomment-3411116656

We also had to re-work our airflow-3 package to ensure that we are
respecting the constraints so that we don't bump package versions to an
unsupported version.
More information here:
https://github.com/wolfi-dev/os/pull/69508

Signed-off-by: David Negreira <david.negreira@chainguard.dev>