dotnet-bootstrap-8/dotnet-bootstrap-9: re-write advisory (#24368)
The advisory for the dotnet-bootstrap-8 and dotnet-bootstrap-9 for CVE-2025-55315
was confusing.
This is an attempt to improve the wording and improve the explanation of
why we do ship a lower version of .NET in the dotnet-bootstrap packages.

Follow-up of https://github.com/wolfi-dev/advisories/pull/24287 which
has more context as well.

Signed-off-by: David Negreira <david.negreira@chainguard.dev>