renovate: add advisories (#25235)
Add advisories for the golang CVEs:
* CVE-2025-47912
* CVE-2025-58183
* CVE-2025-58185
* CVE-2025-58186
* CVE-2025-58187
* CVE-2025-58189
* CVE-2025-58188
* CVE-2025-61723
* CVE-2025-61724
* CVE-2025-61725

This CVE is detected due to a transient dependency of esbuild v0.25.11.
esbuild 0.25.11 is still built with go1.23.12, we will have to wait for esbuild to release a new version with a go version >=1.25.3 or >=1.24.9 and for renovate to release a new version with a new esbuild version.

This should be fixed soon as esbuild has an upstream PR to bump their
golang version: https://github.com/evanw/esbuild/pull/4318

Signed-off-by: David Negreira <david.negreira@chainguard.dev>