renovate: update advisory
Update advisory for CVE-2025-64118
The tar in the renovate package is brought in via a transient dependency
of npm 11.6.2.
Currently npm still pins tar to 7.5.1 in their latest version, we will
have to wait for npm to bump their dependency and renovate to pin npm to
the new npm version after it is released.

Renovate lock:
https://github.com/renovatebot/renovate/blob/f4b569a5959216e5132876a8a8756798750c1ec0/pnpm-lock.yaml#L12081
tar dependency in npm:
https://github.com/npm/cli/blob/072253549d774893a3689341dbc660cb845ebcfe/package.json#L113

Signed-off-by: David Negreira <david.negreira@chainguard.dev>