falcoctl: update advisories (#27905)
Update advisories for CVE-2025-66564 and CVE-2025-66506

github.com/sigstore/fulcio and github.com/sigstore/timestamp-authority
are transitive dependencies pulled in by github.com/sigstore/cosign.

We have tried to bump the transitive dependencies in falcoctl
without success, we need to wait for the migration to sigstore v3 to
happen in falcoctl in order to fix these CVEs.

The bump has already happened in the upstream sigstore v3.0.3 version. [1]

[1] https://github.com/sigstore/cosign/commit/5a60384c2797eaef2520ccbcf7c5af8e97a727f0

Signed-off-by: David Negreira <david.negreira@chainguard.dev>